refactor: PopupComponents를 designsystem으로 옮김
