+++ core/designsystem/src/main/java/com/practice/designsystem/components/PopupComponents.kt
@@ -1,4 +1,4 @@
-package com.practice.main.popup
+package com.practice.designsystem.components
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.padding
